switch to balance model in ui
