+++ ui/src/main/resources/templates/channel-details.html
@@ -10,7 +10,7 @@
 
     <div th:replace="~{fragments/tiles :: channel-details(${channel})}"/>
     <div th:replace="~{fragments/tiles :: fee-policy(${channel.channelStatus}, ${channel.policies})}"/>
-    <div th:replace="~{fragments/tiles :: balance(${channel.balanceInformation}, ${channel.getInboundPercentage}, ${channel.getOutboundPercentage}, ${channel.routableCapacity == 0})}"/>
+    <div th:replace="~{fragments/tiles :: balance(${channel.balanceInformation})}"/>
 
     <div th:replace="~{fragments/tiles :: on-chain-costs(${channel.onChainCosts})}"/>
     <div th:replace="~{fragments/tiles :: rebalance-report(${channel.rebalanceReport})}"/>
