+++ ui/src/main/resources/templates/fragments/tiles.html
@@ -149,9 +149,8 @@
     </div>
 </div>
 
-
-<div th:fragment="balance (balanceInformation, inboundPercent, outboundPercent, disabled)" class="tile col-sm-4"
-     th:if="!${disabled}">
+<div th:fragment="balance (balanceInformation)" class="tile col-sm-4"
+     th:if="${balanceInformation.routableCapacity > 0}">
     <div class="card border-secondary">
         <div class="card-header">
             <div class="fs-4 mb-0">
@@ -162,12 +161,15 @@
         <div class="card-body">
             <div class="container" style="padding-left: 0; padding-right: 0;">
                 <div class="progress" style="margin-bottom: 16px">
-                    <div th:text="${#numbers.formatDecimal(outboundPercent,1,0)} + '%'"
+                    <div th:text="${#numbers.formatDecimal(balanceInformation.getOutboundPercentage,1,0)} + '%'"
                          class="progress-bar outbound-green" role="progressbar"
-                         th:style="'width: ' + ${outboundPercent} + '%'" aria-valuenow="30" aria-valuemin="0"
+                         th:style="'width: ' + ${balanceInformation.getOutboundPercentage} + '%'" aria-valuenow="30"
+                         aria-valuemin="0"
                          aria-valuemax="100"></div>
-                    <div th:text="${#numbers.formatDecimal(inboundPercent,1,0)} + '%'" class="progress-bar inbound-red"
-                         role="progressbar" th:style="'width: ' + ${inboundPercent} + '%'" aria-valuenow="15"
+                    <div th:text="${#numbers.formatDecimal(balanceInformation.getInboundPercentage,1,0)} + '%'"
+                         class="progress-bar inbound-red"
+                         role="progressbar" th:style="'width: ' + ${balanceInformation.getInboundPercentage} + '%'"
+                         aria-valuenow="15"
                          aria-valuemin="0" aria-valuemax="100"></div>
                 </div>
                 <div class="row align-items-start">
