+++ ui/src/test/java/de/cotto/lndmanagej/ui/dto/ChannelDetailsDtoTest.java
@@ -49,21 +49,6 @@ class ChannelDetailsDtoTest {
     }
 
     @Test
-    void routableCapacity() {
-        assertThat(CHANNEL_DETAILS_DTO.getRoutableCapacity()).isEqualTo(1123L);
-    }
-
-    @Test
-    void inboundPercentage() {
-        assertThat(CHANNEL_DETAILS_DTO.getInboundPercentage()).isEqualTo(10.952_804_986_642_917);
-    }
-
-    @Test
-    void outboundPercentage() {
-        assertThat(CHANNEL_DETAILS_DTO.getOutboundPercentage()).isEqualTo(100 - 10.952_804_986_642_917);
-    }
-
-    @Test
     void capacitySat() {
         assertThat(CHANNEL_DETAILS_DTO.capacitySat()).isEqualTo(21_000_000);
     }
